refactor: improve code quality and request typing and remove unsafe casts

+++ src/modules/products/ProductController.ts
@@ -29,11 +29,18 @@ class ProductController {
   }
   async deleteProduct(req: Request, res: Response): Promise<Response> {
     let productId = Number(req.params.productId);
+    const userId = req.userId;
+    if (!userId) {
+      throw new AppError(401, "Not Authenticated");
+    }
     await this.service.deleteProduct(productId);
     return res.status(204).send();
   }
-  async listProducts(req: Request, res: Response): Promise<Response> {
-    const { limit, offset } = req.query as unknown as GetProductsQueryInput;
+  async listProducts(
+    req: Request<{}, {}, {}, GetProductsQueryInput>,
+    res: Response,
+  ): Promise<Response> {
+    const { limit, offset } = req.query;
     const products = await this.service.listProducts(limit, offset);
     return res.status(200).json(products);
   }
